rebase tradermade changes
